feat(runtime) Use pre-compiled dynamic libraries.
This patch depends on https://github.com/wasmerio/wasmer/pull/474,
which fixes the dylib ID for macOS to use `@rpath`.

This patch renames the `just go` recipe to `just build`.

The `just build` recipe compiles the Rust dynamic library only when
required, i.e. when the `libwasmer_runtime_c_api.*` file is absent
from the source tree.

This patch also updates `bridge.go` to specify the `rpath` to the
linker with `-Wl,-rpath`. On macOS, the solution would be to set
`rpath` to `@executable_path` but since
https://nvd.nist.gov/vuln/detail/CVE-2018-6574, some flags are
disabled (see https://github.com/golang/go/wiki/InvalidFlag for more
information). Hopefully for us, cgo has this special `${SRCDIR}`
variable (see
https://golang.org/cmd/cgo/#hdr-Using_cgo_with_the_go_command), which
is a possible alternative.

+++ justfile
@@ -1,9 +1,5 @@
-# Compile the Rust part for this specific system.
-rust:
-	cargo build --release
-
-# Compile the Go part.
-go go-build-args='':
+# Compile the library.
+build go-build-args='-v':
 	#!/usr/bin/env bash
 	set -euo pipefail
 	cd wasmer
@@ -17,9 +13,15 @@ go go-build-args='':
 		*)
 			dylib_extension="so"
 	esac
-	test -f libwasmer_runtime_c_api.${dylib_extension} || \
+	if ! test -f libwasmer_runtime_c_api.${dylib_extension}; then
+		cargo build --release
 		ln -s ../target/release/deps/libwasmer_runtime_c_api-*.${dylib_extension} libwasmer_runtime_c_api.${dylib_extension}
-	go build -ldflags="-r $(pwd)" -v {{go-build-args}} .
+	fi
+	go build {{go-build-args}} .
+
+# Compile the Rust part for this specific system.
+rust:
+	cargo build --release
 
 # Generate cgo debug objects.
 debug-cgo:
@@ -30,7 +32,7 @@ debug-cgo:
 # Run all the tests.
 test:
 	#!/usr/bin/env bash
-	export LD_LIBRARY_PATH=$(pwd)/wasmer
+	#export DYLD_PRINT_LIBRARIES=y
 	cd wasmer/test/
 	# Run the tests.
 	go test -test.v $(find . -type f \( -name "*_test.go" \! -name "example_*.go" \! -name "benchmark*.go" \) ) imports.go
